refactor(tkinter): rename UI entry modules and add compatibility launchers
