+++ README.de.md
@@ -24,7 +24,7 @@ KEYWORDS: `EPUB Uebersetzer`, `EPUB Uebersetzungswerkzeug`, `KI Uebersetzung`, `
 ### Tkinter
 ```powershell
 cd project-tkinter
-python start.py
+python app_main.py --variant classic
 ```
 
 ### Web Desktop
