+++ legacy/README.md
@@ -5,7 +5,8 @@ This folder contains archived root-level scripts from an older repository layout
 - `tlumacz_ollama.py`
 
 Use current project entry points instead:
-- `project-tkinter/start.py`
+- `project-tkinter/app_main.py` (`--variant classic|horizon`)
+- `project-tkinter/start.py` (compat alias)
 - `project-web-desktop/run-backend.ps1`
 - `project-web-desktop/run-desktop.ps1`
